+++ src/main/java/machir/emitters/client/ClientProxy.java
@@ -15,7 +15,7 @@ import net.minecraft.client.particle.EntitySplashFX;
 
 public class ClientProxy extends CommonProxy {
 	@Override
-	public void spawnParticle(TileEmitter emitter, float xOffset, float yOffset, float zOffset, float xVel, float yVel, float zVel) {
+	public void spawnParticle(TileEmitter emitter, double xOffset, double yOffset, double zOffset, double xVel, double yVel, double zVel) {
     	// Switch over the different particle types and create the correct one
 		EntityFX emitterFX = null;
     	switch(emitter.getParticle()) {
@@ -26,22 +26,22 @@ public class ClientProxy extends CommonProxy {
     		emitterFX = new EntityFlameFX(emitter.getWorldObj(), emitter.xCoord + xOffset, emitter.yCoord + yOffset, emitter.zCoord + zOffset, xVel, yVel, zVel);
     		break;
     	case HEART:
-    		emitterFX = new EntityHeartFX(emitter.getWorldObj(), emitter.xCoord + xOffset, emitter.yCoord + yOffset, emitter.zCoord + zOffset, xVel, yVel, zVel, emitter.size);
+    		emitterFX = new EntityHeartFX(emitter.getWorldObj(), emitter.xCoord + xOffset, emitter.yCoord + yOffset, emitter.zCoord + zOffset, xVel, yVel, zVel, (float)emitter.size);
     		break;
     	case LAVA:
     		emitterFX = new EntityLavaFX(emitter.getWorldObj(), emitter.xCoord + xOffset, emitter.yCoord + yOffset, emitter.zCoord + zOffset);
     		break;
     	case NOTE:
-    		emitterFX = new EntityNoteFX(emitter.getWorldObj(), emitter.xCoord + xOffset, emitter.yCoord + yOffset, emitter.zCoord + zOffset, xVel, yVel, zVel, emitter.size);
+    		emitterFX = new EntityNoteFX(emitter.getWorldObj(), emitter.xCoord + xOffset, emitter.yCoord + yOffset, emitter.zCoord + zOffset, xVel, yVel, zVel, (float)emitter.size);
     		break;
     	case SPLASH:
     		emitterFX = new EntitySplashFX(emitter.getWorldObj(), emitter.xCoord + xOffset, emitter.yCoord + yOffset, emitter.zCoord + zOffset, xVel, yVel, zVel);
     		break;
     	case REDDUST:
-    		emitterFX = new EntityReddustFX(emitter.getWorldObj(), emitter.xCoord + xOffset, emitter.yCoord + yOffset, emitter.zCoord + zOffset, emitter.size, xVel, yVel, zVel);
+    		emitterFX = new EntityReddustFX(emitter.getWorldObj(), emitter.xCoord + xOffset, emitter.yCoord + yOffset, emitter.zCoord + zOffset, (float)emitter.size, (float)xVel, (float)yVel, (float)zVel);
     		break;
     	case SMOKE:
-    		emitterFX = new EntitySmokeFX(emitter.getWorldObj(), emitter.xCoord + xOffset, emitter.yCoord + yOffset, emitter.zCoord + zOffset, xVel, yVel, zVel, emitter.size);
+    		emitterFX = new EntitySmokeFX(emitter.getWorldObj(), emitter.xCoord + xOffset, emitter.yCoord + yOffset, emitter.zCoord + zOffset, xVel, yVel, zVel, (float)emitter.size);
     		break;
     	default:
 			return;
